Adjust types
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1395 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/residfp-builder/residfp/WaveformGenerator.cpp
@@ -113,7 +113,7 @@ void WaveformGenerator::setChipModel(ChipModel chipModel)
         dac[i] = (short)(dacValue + 0.5);
     }
 
-    const int offset = dac[chipModel == MOS6581 ? 0x380 : 0x800];
+    const short offset = dac[chipModel == MOS6581 ? 0x380 : 0x800];
 
     for (int i = 0; i < (1 << DAC_BITS); i ++)
     {
